Mere clean version, får stadig  Column 'student_usernames' cannot be null og den indlæser ikke CourseID når jeg getter

+++ lesson-app/src/main/java/dk/aau/cs/ds302e18/app/domain/Course.java
@@ -1,14 +1,14 @@
 package dk.aau.cs.ds302e18.app.domain;
 
 public class Course {
-    private Integer courseID;
+    private long courseID;
     private String studentUsernames;
 
-    public int getCourseID() {
+    public long getCourseID() {
         return courseID;
     }
 
-    public void setCourseID(int courseID) {
+    public void setCourseID(long courseID) {
         this.courseID = courseID;
     }
